adds -c argument
Adds ability to have local and global session files for easy sharing.
Local session files don't need to cd anywhere while global sessions do.
this way one can share a tmuxstart file with a project, or not.
while calling tmuxstart it will check if the session file is a symlink
and it will cd into it's directory before sourcing it.